build: remove esbuild
