--- .esbuildrc.js
@@ -1,37 +1,0 @@
-import fs from 'fs';
-import { resolve } from 'path';
-import { build } from 'esbuild';
-
-function read() {
-  const PACKAGES_DIRECTORY = 'packages';
-  const packages = fs.readdirSync(PACKAGES_DIRECTORY);
-  return packages.map((pack) => {
-    const root = `${PACKAGES_DIRECTORY}/${pack}`;
-    const main = resolve(`${root}/lib/index.cjs`);
-    const src = resolve(`${root}/src`);
-    return {
-      main,
-      src,
-    };
-  });
-}
-
-if (process.env.npm_lifecycle_event === 'build') {
-  read().forEach((pack) => {
-    build({
-      allowOverwrite: true,
-      bundle: true,
-      entryPoints: fs.readdirSync(pack.src)
-        .filter((file) => file.endsWith('.cjs'))
-        .map((file) => `${pack.src}/${file}`),
-      outfile: pack.main,
-      platform: 'node',
-    }).catch((error) => {
-      throw new Error(error.stderr);
-    });
-  });
-}
-
-export {
-  read,
-};
