+++ .github/workflows/build.yml
@@ -22,6 +22,5 @@ jobs:
           >> .npmrc
       - run: npm i -g pnpm
       - run: pnpm i
-      - run: pnpm build
       - run: pnpm lint
       - run: pnpm test
